LPS-123094 Extract Editor component to avoid unnecessary render
